Simplify request data handling and improve token naming
Refactored the way request data is handled in POST requests to increase code readability. The unnecessary additional step of assigning request POST data to a variable before processing it was removed. In addition, renamed 'usertoken_set.first()' to 'usertoken_set.first().token' for better semantic clarity. This change correctly reflects that the actual token, not the entire UserToken object, is returned in the response. These changes aim to improve the structure of the code for better future maintainability.