"Improve authentication and update libraries"
This set of changes is primarily focused on improving the authentication process of the system and updating required libraries.

In views.py file, all the functions have been converted to comments, and an email field is added in the User model. Whereas, in settings.py, "rest_framework.permissions.IsAuthenticated" is set to default permission class, and "rest_framework_simplejwt.authentication.JWTAuthentication" to default authentication class. The UserToken model in models.py has been removed. Moreover, changes in requirements.txt reflect the update of versions of various libraries and the replacement of psycopg2-binary with mysqlclient.

The diff also includes adjustments in url routing and a major rewrite of endpoints.py, incorporating new classes like CheckAuthenticatedView, GetCSRFToken, RegisterView, LoginView, LogoutView for handling authentication related actions. The most important change is the functional adjustments in the authentication process, specifically the implementation of JSON Web Tokens for stateless, secure data transfer.

The authentication changes will improve the security and efficiency of the system while updated libraries offer smoother functioning and new features for use in the project.